Making progress through the boot sequence. Implementing more missing opcodes.

+++ Machines/Zexdoc.cs
@@ -26,6 +26,6 @@ public class Zexdoc : IMemoryProvider
     {
         Array.Clear(Memory, 0, Memory.Length);
     }
-    public byte ReadPort(byte port) => 0xFF; // Zexdoc doesn't use ports
-    public void WritePort(byte port, byte value) { }
+    public byte ReadPort(ushort port) => 0xFF; // Zexdoc doesn't use ports
+    public void WritePort(ushort port, byte value) { }
 }
